Refactor the API into multiple modules
