+++ api.go
@@ -1,0 +1,7 @@
+package wunderground
+
+type Response struct {
+	Version string `json:"version,omitempty"`
+	TOS     string `json:"termsOfService,omitempty"`
+}
+
